+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import App from 'App';
 import { Provider } from 'react-redux';
-import { store } from 'common/store';
+import { store } from 'common/store/store';
 import 'reset.css';
 import 'index.css';
